refactor(): WIP - Rename server settings
