+++ settings/dev-settings.yaml
@@ -1,29 +1,27 @@
-Environment : "development"
-
 Server:
-  Port : 8000
-  Host : "localhost"
-  Domain : "http://localhost:8000"
+  Port: 8000
+  Host: "localhost"
+  Domain: "http://localhost:8000"
 
 RepositorySettings:
-  Host : localhost
-  Port : 3306
-  Name : "sprinter_db"
-  User : "root"
-  Password : "root"
+  Host: localhost
+  Port: 3306
+  Name: "sprinter_db"
+  User: "root"
+  Password: "root"
 
 LogSettings:
-  LogDir : "./runtime/logs"
+  LogDir: "./runtime/logs"
 
 Paseto:
-  PasetoSecurityKey : "b7e16bf2-ab0f-4a68-a16b-f2ab0f6a"
+  PasetoSecurityKey: "b7e16bf2-ab0f-4a68-a16b-f2ab0f6a"
 
 FileStorage:
-  StorageFolder : "./runtime/storage"
+  StorageFolder: "./runtime/storage"
 
 SMTPConfig:
-  Host : ""
-  Port : ""
-  User : ""
-  Password : ""
-  From : ""
+  Host: ""
+  Port: ""
+  User: ""
+  Password: ""
+  From: ""
